server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.15.0
supabase==2.11.0
sentence-transformers==3.3.1
python-dotenv==1.0.1
httpx==0.28.1
numpy==1.26.4
